chore: move safeload_yaml tests
